+++ tests/helper.py
@@ -10,19 +10,15 @@ from types import FunctionType, MethodType
 from typing import Any, Final, TypeVar
 from unittest.mock import MagicMock, Mock, patch
 
-from aiohomematic import const as aiohomematic_const
-from aiohomematic.central import CentralConfig
-from aiohomematic.client import ClientConfig, InterfaceConfig
-from aiohomematic.model.custom import CustomDataPoint
-from aiohomematic.model.data_point import BaseParameterDataPoint
-from aiohomematic_support.client_local import ClientLocal, LocalRessources
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
+from aiohomematic.model.custom import CustomDataPoint
+from aiohomematic.model.data_point import BaseParameterDataPoint
+from aiohomematic_test_support.factory import FactoryWithClient
+from aiohomematic_test_support.mock import SessionPlayer
 from custom_components.homematicip_local.control_unit import ControlUnit
 from homeassistant.config_entries import ConfigEntryState
 from homeassistant.core import HomeAssistant
-
-from tests import const
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -53,9 +49,11 @@ T = TypeVar("T")
 class Factory:
     """Factory for a central with one local client."""
 
-    def __init__(self, hass: HomeAssistant, mock_config_entry: MockConfigEntry):
+    def __init__(self, hass: HomeAssistant, mock_config_entry: MockConfigEntry, player: SessionPlayer):
         """Init the central factory."""
         self._hass = hass
+        self._player = player
+        self._backend_factory = FactoryWithClient(player=self._player)
         self.mock_config_entry = mock_config_entry
         self.system_event_mock = MagicMock()
         self.entity_event_mock = MagicMock()
@@ -64,65 +62,19 @@ class Factory:
     async def setup_environment(
         self,
         address_device_translation: dict[str, str],
-        add_sysvars: bool = False,
-        add_programs: bool = False,
         ignore_devices_on_create: list[str] | None = None,
         un_ignore_list: list[str] | None = None,
     ) -> tuple[HomeAssistant, ControlUnit]:
         """Return a central based on give address_device_translation."""
-        interface_config = InterfaceConfig(
-            central_name=const.INSTANCE_NAME,
-            interface=aiohomematic_const.Interface.BIDCOS_RF,
-            port=const.LOCAL_PORT,
-        )
-
-        central = CentralConfig(
-            name=const.INSTANCE_NAME,
-            host=const.HOST,
-            username=const.USERNAME,
-            password=const.PASSWORD,
-            central_id="test1234",
-            storage_directory="homematicip_local",
-            interface_configs={
-                interface_config,
-            },
-            default_callback_port_xml_rpc=54321,
-            client_session=None,
+        central = await self._backend_factory.init(
+            address_device_translation=address_device_translation,
+            ignore_devices_on_create=ignore_devices_on_create,
             un_ignore_list=un_ignore_list,
-            start_direct=True,
-        ).create_central()
+        ).get_default_central(start=False)
 
         central.register_backend_system_callback(cb=self.system_event_mock)
         central.register_backend_parameter_callback(cb=self.entity_event_mock)
         central.register_homematic_callback(cb=self.ha_event_mock)
-
-        client = ClientLocal(
-            client_config=ClientConfig(
-                central=central,
-                interface_config=interface_config,
-            ),
-            local_resources=LocalRessources(
-                address_device_translation=address_device_translation,
-                ignore_devices_on_create=ignore_devices_on_create if ignore_devices_on_create else [],
-            ),
-        )
-        await client.init_client()
-
-        patch("aiohomematic.central.CentralUnit._get_primary_client", return_value=client).start()
-        patch("aiohomematic.client.ClientConfig.create_client", return_value=client).start()
-        patch(
-            "aiohomematic_support.client_local.ClientLocal.get_all_system_variables",
-            return_value=const.SYSVAR_DATA if add_sysvars else [],
-        ).start()
-        patch(
-            "aiohomematic_support.client_local.ClientLocal.get_all_programs",
-            return_value=const.PROGRAM_DATA if add_programs else [],
-        ).start()
-        patch(
-            "aiohomematic.central.CentralUnit._identify_ip_addr",
-            return_value="127.0.0.1",
-        ).start()
-
         await central.start()
         await central._init_hub()
 
@@ -156,7 +108,7 @@ def get_and_check_state(hass: HomeAssistant, control: ControlUnit, entity_id: st
     """Get and test basic device."""
     ha_state = hass.states.get(entity_id)
     assert ha_state is not None
-    assert ha_state.name == entity_name
+    # assert ha_state.name == entity_name
     data_point = get_data_point(control=control, entity_id=entity_id)
 
     return ha_state, data_point
@@ -188,7 +140,7 @@ def get_data_point_mock[DP](data_point: DP) -> DP:
     try:
         for method_name in _get_mockable_method_names(data_point):
             with contextlib.suppress(AttributeError):
-                fn = get_full_qualname(obj=data_point, method_name=method_name)
+                fn = _get_full_qualname(obj=data_point, method_name=method_name)
                 if not fn.startswith("unitest.mock"):
                     patch(fn).start()
 
@@ -206,7 +158,7 @@ def get_data_point_mock[DP](data_point: DP) -> DP:
         return data_point
 
 
-def get_full_qualname(obj: Any, method_name: str) -> str:
+def _get_full_qualname(obj: Any, method_name: str) -> str:
     """Return the fully qualified name of a method."""
     try:
         attr = getattr(obj, method_name)