+++ tests/test_sensor.py
@@ -7,6 +7,7 @@ import pytest
 from homeassistant.const import STATE_UNKNOWN
 
 from tests import const, helper
+from tests.helper import Factory
 
 TEST_DEVICES: dict[str, str] = {
     "VCU7837366": "HB-UNI-Sensor1.json",
@@ -16,12 +17,12 @@ TEST_DEVICES: dict[str, str] = {
 
 
 @pytest.mark.asyncio
-async def test_sensor_trans(factory: helper.Factory) -> None:
+async def test_sensor_trans(factory_homegear: Factory) -> None:
     """Test sensor with translation."""
     entity_id = "sensor.hb_uni_sensor1_vcu7837366_dew_point"
     entity_name = "HB-UNI-Sensor1_VCU7837366 dew point"
 
-    hass, control = await factory.setup_environment(TEST_DEVICES)
+    hass, control = await factory_homegear.setup_environment(TEST_DEVICES)
     ha_state, data_point = helper.get_and_check_state(
         hass=hass, control=control, entity_id=entity_id, entity_name=entity_name
     )
@@ -41,12 +42,12 @@ async def test_sensor_trans(factory: helper.Factory) -> None:
 
 
 @pytest.mark.asyncio
-async def test_sensor_to_trans(factory: helper.Factory) -> None:
+async def test_sensor_to_trans(factory_homegear: Factory) -> None:
     """Test sensor without translation."""
     entity_id = "sensor.hb_uni_sensor1_vcu7837366_abs_luftfeuchte"
     entity_name = "HB-UNI-Sensor1_VCU7837366 Abs Luftfeuchte"
 
-    hass, control = await factory.setup_environment(TEST_DEVICES)
+    hass, control = await factory_homegear.setup_environment(TEST_DEVICES)
     ha_state, data_point = helper.get_and_check_state(
         hass=hass, control=control, entity_id=entity_id, entity_name=entity_name
     )